aaaaaa

+++ src/com/internousdev/freesia/action/AdminItemUpdateAction.java
@@ -52,55 +52,32 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
      */
     private static final long serialVersionUID = -2457344252220177410L;
 
-    /**
-     * リザルト
-     */
-    private String result;
 
     /**
      * 商品ID
      */
-    private int itemId;
+    private int userId;
 
     /**
-     * カテゴリーID
+     * 画像ファイル
      */
-    private int categoryId;
+    private File tmpfile;
 
     /**
      * 商品名
      */
-    private String itemName;
+    private String herName;
 
-    /**
-     * 単価
-     */
-    private float price;
+    private String title;
 
-    /**
-     * 在庫数
-     */
-    private int stocks;
-
-    /**
-     * 商品詳細
-     */
-    private String itemDetail;
+    private String comment;
 
     /**
      * イメージパス
      */
-    private String[] imagePath={"","","","",""};
+    private String imagepath1;
 
-    /**
-     * プレビューパス
-     */
-    private String[] prevPath;
 
-    /**
-     * 画像削除情報
-     */
-    private int[] imageDelete={-50};
 
     /**
      * アップロード情報
@@ -108,19 +85,10 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
     private boolean[] uploaded;
 
     /**
-     * 画像配列
-     */
-    private File[] image={new File("nullfile")};
-
-    /**
      * 表示画像
      */
     private BufferedImage imageData;
 
-    /**
-     * 画像ファイル
-     */
-    private File tmpfile;
 
 
     /**
@@ -128,95 +96,59 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
     * @author JUNNYA TAKENO
     * @since 2017/07/16
     * @return 成功ならSUCCESS それ以外はERROR
+
     */
-    public String execute() throws Exception{
-        result = ERROR;
-        AdminItemUpdateDAO dao= new AdminItemUpdateDAO();
+    //public String execute() throws Exception{
+        //result = ERROR;
+        //AdminItemUpdateDAO dao= new AdminItemUpdateDAO();
 
-        int i=0;
-        int imagesetcount=0;
-        int uploadcount=0;
-        int deletecount=0;
+      //  int i=0;
+      //  int imagesetcount=0;
+     //   int uploadcount=0;
+     //   int deletecount=0;
 
-        String basePath=request.getServletContext().getRealPath("/");
-
-        for(i=0;i<imagePath.length;i++){
+     //   String basePath=request.getServletContext().getRealPath("/");
 
 
-            if(imageDelete[0]>=0&&deletecount<imageDelete.length&&i==imageDelete[deletecount]){
-                if(uploaded[i]){
-                    uploadcount++;
-                }
-                deletecount++;
-                continue;
-            }else if(uploaded[i]){
-                imagePath[imagesetcount]=uploadFile(imagesetcount,uploadcount,basePath);
-                uploadcount++;
-                if(imagePath[imagesetcount]=="failed"){
-                    imagePath[imagesetcount]="";
-                    continue;
-                }else{
-                    imagesetcount++;
-                }
-            }else if(prevPath[i].length()>0){
-                imagePath[imagesetcount]=renameFile(imagesetcount,i,prevPath[i],basePath);
-                if(imagePath[imagesetcount]=="failed"){
-                    imagePath[imagesetcount]="";
-                    continue;
-                }else{
-                    imagesetcount++;
-                }
-            }
-        }
+        //result = dao.insert(herId,herName,herDetail,imagepath1);
+        //session.put("adminEditItemId", 0);
+        //session.remove("adminEditItemData");
+        //session.put("adminItemLoad", false);
 
-        result = dao.update(itemId,categoryId,itemName,price,stocks,itemDetail,imagePath);
-        session.put("adminEditItemId", 0);
-        session.remove("adminEditItemData");
-        session.put("adminItemLoad", false);
+       // result = SUCCESS;
 
-        result = SUCCESS;
-
-        return result;
-    }//execute
+     //   return result;
+   // }//execute
     //class
 
+    public String execute() throws Exception {
 
-    /**
-    * 商品画像を追加する実行メソッド
-    * @param picno 商品画像番号
-    * @param num 送信画像番号
-    * @param basePath 送信元画像パス
-    * @return 成功なら画像パス それ以外はERROR
-    * @throws Exception 例外
-    */
-    public String uploadFile(int picno,int num,String basePath) throws Exception{
-        String lastPath="img/items/ItemImage_"+itemId+"_"+(picno+1)+".png";
-        String fileName=basePath+lastPath;
+        String result = ERROR;
+        AdminItemUpdateDAO dao = new AdminItemUpdateDAO();
 
-        tmpfile=new File(fileName);
-        tmpfile.delete();
-        tmpfile.createNewFile();
 
-        imageData=ImageIO.read(image[num]);
-
-        if(ImageIO.write(imageData, "png", tmpfile)){
-            return lastPath;
-        }else{
-            return "failed";
+        if (session.get("userId") != null) {
+            userId = (int) session.get("userId");
+            result= dao.insert(userId, herName, comment, title, imagepath1);
         }
+        result = SUCCESS;
+
+
+        return result;
     }
 
 
-    /**
-    * する実行メソッド
-    * @param picno 商品画像番号
-    * @param num 送信画像番号
-    * @param prevPath プレビューパス
-    * @param basePath 送信元画像パス
-    * @return 成功なら画像パス それ以外はfailed
-    */
-    public String renameFile(int picno,int num,String prevPath,String basePath){
-        String lastPath="img/items/ItemImage_"+itemId+"_"+(picno+1)+".png";
+    public String uploadFile(int picno,int num,String basePath) throws Exception{
+        String lastPath="img/collections"+(picno+1)+".png";
+        String fileName=basePath+lastPath;
+
+        tmpfile=new File(fileName);
+        tmpfile.createNewFile();
+        return lastPath;
+    }
+
+    public String renameFile(int picno,int num,String prevPath,String basePath) throws Exception{
+        String lastPath="img/collections"+(picno+1)+".png";
         String fileName_new=basePath+lastPath;
         String fileName_old=basePath+prevPath;
         String fileName_copy=fileName_old+"_copy.png";
@@ -226,20 +158,26 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
         File tmpfile_copy=new File(fileName_copy);
         tmpfile_copy.delete();
 
-        tmpfile_old.renameTo(tmpfile_copy);
-        tmpfile_new.delete();
-        tmpfile_old.delete();
-
-        if(tmpfile_copy.renameTo(tmpfile_new)){
-            tmpfile_copy.delete();
-            return lastPath;
+        if(fileName_old.indexOf("/ItemImage_")!=-1){
+            tmpfile_old.renameTo(tmpfile_copy);
+            tmpfile_new.delete();
+            tmpfile_old.delete();
+            if(tmpfile_copy.renameTo(tmpfile_new)){
+                tmpfile_copy.delete();
+                return lastPath;
+            }else{
+                return "failed";
+            }
         }else{
-            return "failed";
+            tmpfile_new.createNewFile();
+            imageData=ImageIO.read(tmpfile_old);
+            if(ImageIO.write(imageData, "png", tmpfile_new)){
+                return lastPath;
+            }else{
+                return "failed";
+            }
         }
     }
-
-
-
     /**
      * セッション情報を取得
      * @return session セッション情報
@@ -260,145 +198,68 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
      * 商品IDを取得
      * @return itemId 商品ID
      */
-    public int getItemId() {
-        return itemId;
+    public int getUserId() {
+        return userId;
     }
 
     /**
      * 商品IDを設定
      * @param itemId 商品ID
      */
-    public void setItemId(int itemId) {
-        this.itemId = itemId;
-    }
-
-    /**
-     * カテゴリーIDを取得
-     * @return categoryId カテゴリーID
-     */
-    public int getCategoryId() {
-        return categoryId;
-    }
-
-    /**
-     * カテゴリーIDを設定
-     * @param categoryId カテゴリーID
-     */
-    public void setCategoryId(int categoryId) {
-        this.categoryId = categoryId;
+    public void setUserId(int userId) {
+        this.userId = userId;
     }
 
     /**
      * 商品名を取得
      * @return itemName 商品名
      */
-    public String getItemName() {
-        return itemName;
+    public String getHerName() {
+        return herName;
     }
 
     /**
      * 商品名を設定
      * @param itemName 商品名
      */
-    public void setItemName(String itemName) {
-        this.itemName = itemName;
+    public void setHerName(String herName) {
+        this.herName = herName;
     }
 
-    /**
-     * 単価を取得
-     * @return price 単価
-     */
-    public float getPrice() {
-        return price;
-    }
-
-    /**
-     * 単価を設定
-     * @param price 単価
-     */
-    public void setPrice(float price) {
-        this.price = price;
-    }
-
-    /**
-     * 在庫数を取得
-     * @return stocks 在庫数
-     */
-    public int getStocks() {
-        return stocks;
-    }
-
-    /**
-     * 在庫数を設定
-     * @param stocks 在庫数
-     */
-    public void setStocks(int stocks) {
-        this.stocks = stocks;
-    }
 
     /**
      * 商品詳細を取得
      * @return itemDetail 商品詳細
      */
-    public String getItemDetail() {
-        return itemDetail;
+    public String getComment() {
+        return comment;
     }
 
     /**
      * 商品詳細を設定
      * @param itemDetail 商品詳細
      */
-    public void setItemDetail(String itemDetail) {
-        this.itemDetail = itemDetail;
+    public void setComment(String comment) {
+        this.comment = comment;
     }
 
     /**
      * イメージパスを取得
      * @return imagePath イメージパス
      */
-    public String[] getImagePath() {
-        return imagePath;
+    public String getImagepath1() {
+        return imagepath1;
     }
 
     /**
      * イメージパスを設定
      * @param imagePath イメージパス
      */
-    public void setImagePath(String[] imagePath) {
-        this.imagePath = imagePath;
+    public void setImagepath1(String imagepath1) {
+        this.imagepath1 = imagepath1;
     }
 
-    /**
-     * プレビューパスを取得
-     * @return prevPath プレビューパス
-     */
-    public String[] getPrevPath() {
-        return prevPath;
-    }
 
-    /**
-     * プレビューパスを設定
-     * @param prevPath プレビューパス
-     */
-    public void setPrevPath(String[] prevPath) {
-        this.prevPath = prevPath;
-    }
-
-    /**
-     * 画像削除情報を取得
-     * @return imageDelete 画像削除情報
-     */
-    public int[] getImageDelete() {
-        return imageDelete;
-    }
-
-    /**
-     * 画像削除情報を設定
-     * @param imageDelete 画像削除情報
-     */
-    public void setImageDelete(int[] imageDelete) {
-        this.imageDelete = imageDelete;
-    }
 
     /**
      * アップロード情報を取得
@@ -416,21 +277,6 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
         this.uploaded = uploaded;
     }
 
-    /**
-     * 画像配列を取得
-     * @return image 画像配列
-     */
-    public File[] getImage() {
-        return image;
-    }
-
-    /**
-     * 画像配列を設定
-     * @param image 画像配列
-     */
-    public void setImage(File[] image) {
-        this.image = image;
-    }
 
     /**
      * リクエストを取得
@@ -463,4 +309,14 @@ implements SessionAware,ServletRequestAware,ServletResponseAware {
     public void setServletResponse(HttpServletResponse response){
         this.response = response;
     }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+
 }
